extract API key auth into SRP core microcrate
